requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-autoslug-iplweb==1.9.4
django-dashing==0.3.3
django-heroku==0.3.1
django-nine==0.1.13
django-tinymce==2.7.0
dj-database-url==0.5.0
easy-thumbnails==2.5
feedparser==5.2.1
gunicorn==19.9.0
idna==2.7
ordereddict==1.1
pif==0.8.2
Pillow==5.3.0
pip==18.1
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
setuptools==39.0.1
six==1.11.0
transliterate==1.10.2
Unidecode==1.0.22
urllib3==1.24.1
vishap==0.1.5
whitenoise==4.1.2
